requirements.txt
accelerate==1.5.2
anthropic==0.49.0
benepar==0.2.0
bert-score==0.3.13
bitsandbytes==0.45.3
ctc_score==0.1.3
datasets==3.4.0
deepeval==2.5.5
evaluate==0.4.3
hydra-core==1.3.2
nltk==3.9.1
kaleido==0.2.1
omegaconf==2.3.0
openai==1.66.3
openpyxl==3.1.5
peft==0.14.0
plotly==5.23.0
protobuf==3.20.3
pytest==8.3.2
pytorch_lightning==2.5.0.post0
rake_nltk==1.0.6
reportlab==4.3.1
rouge==1.0.1
rouge-score==0.1.2
sacrebleu==2.4.1
spacy==3.7.5
streamlit==1.37.0
streamlit-authenticator==0.4.2
tabulate==0.9.0
transformers==4.49.0
trl==0.15.2
torchmetrics==1.4.1
unsloth==2025.3.17
vllm==0.8.1
xlrd==1.2.0
